requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-filter==24.3